+++ spec/praxis/extensions/attribute_filtering/filters_parser_spec.rb
@@ -118,14 +118,23 @@ describe Praxis::Extensions::AttributeFiltering::FilteringParams::Parser do
         'cOrN.00copia.of_thinGs.42_here=1' => 'cOrN.00copia.of_thinGs.42_here=1',
       }
     end
-    context 'supports everything (except &|(),) for values' do
+    context 'supports everything (except &|(),) for values (even without encoding..not allowed, but just to ensure the parser does not bomb)' do
       it_behaves_like 'round-trip-properly', {
         'v=1123' => 'v=1123',
         'v=*foo*' => 'v=*foo*',
         'v=*^%$#@!foo' => 'v=*^%$#@!foo',
-        'v=_-+=\{}"?:><' => 'v=_-+=\{}"?:><',
-        'v=_-+=\{}"?:><,another_value!' => 'v=[_-+=\{}"?:><,another_value!]',
+        'v=_-=\{}"?:><' => 'v=_-=\{}"?:><',
+        'v=_-=\{}"?:><,another_value!' => 'v=[_-=\{}"?:><,another_value!]',
       }
     end
+    context 'properly handles url-encoded values' do
+      it_behaves_like 'round-trip-properly', {
+        "v=#{CGI.escape('1123')}" => 'v=1123',
+        "v=#{CGI.escape('*foo*')}" => 'v=*foo*',
+        "v=#{CGI.escape('*^%$#@!foo')}" => 'v=*^%$#@!foo',
+        "v=#{CGI.escape('~!@#$%^&*()_+-={}|[]\:";\'<>?,./`')}" => 'v=~!@#$%^&*()_+-={}|[]\:";\'<>?,./`',
+        "v=#{CGI.escape('_-+=\{}"?:><')},#{CGI.escape('another_value!')}" => 'v=[_-+=\{}"?:><,another_value!]',
+      }
+    end 
   end
 end